@@ -26,7 +26,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "Contributors": MessageLookupByLibrary.simpleMessage("Contributors"),
         "Friday": MessageLookupByLibrary.simpleMessage("FRI"),
         "Monday": MessageLookupByLibrary.simpleMessage("MON"),
-        "PrivacyPolicy": MessageLookupByLibrary.simpleMessage("Privacy policy"),
+        "PrivacyPolicy": MessageLookupByLibrary.simpleMessage("Privacy Policy"),
         "Saturday": MessageLookupByLibrary.simpleMessage("SAT"),
         "Sunday": MessageLookupByLibrary.simpleMessage("SUN"),
         "Thursday": MessageLookupByLibrary.simpleMessage("THU"),
@@ -44,6 +44,8 @@ class MessageLookup extends MessageLookupByLibrary {
         "agree": MessageLookupByLibrary.simpleMessage("Agree"),
         "alertError": MessageLookupByLibrary.simpleMessage("An error occurred"),
         "announcement": MessageLookupByLibrary.simpleMessage("Announcements"),
+        "announcementEmpty":
+            MessageLookupByLibrary.simpleMessage("No Announcement"),
         "appWillClose": MessageLookupByLibrary.simpleMessage("APP will close"),
         "app_name": MessageLookupByLibrary.simpleMessage("TAT"),
         "areYouSureToDownload": MessageLookupByLibrary.simpleMessage(
@@ -70,6 +72,8 @@ class MessageLookup extends MessageLookupByLibrary {
         "closeOnce":
             MessageLookupByLibrary.simpleMessage("Press again to close"),
         "contribute": MessageLookupByLibrary.simpleMessage("contribute"),
+        "contributionCourse": MessageLookupByLibrary.simpleMessage(
+            "Contribution to course total"),
         "coreAbility": MessageLookupByLibrary.simpleMessage("Core ability"),
         "course": MessageLookupByLibrary.simpleMessage("Course"),
         "courseContent": MessageLookupByLibrary.simpleMessage("Course content"),
@@ -81,6 +85,10 @@ class MessageLookup extends MessageLookupByLibrary {
         "courseObject": MessageLookupByLibrary.simpleMessage("Course object"),
         "courseRefbook": MessageLookupByLibrary.simpleMessage("Course Refbook"),
         "courseRemark": MessageLookupByLibrary.simpleMessage("Course remark"),
+        "courseSearchHint": MessageLookupByLibrary.simpleMessage(
+            "Enter text above to search for courses"),
+        "courseSearchNotFound":
+            MessageLookupByLibrary.simpleMessage("No relevant courses found"),
         "courseTextbook":
             MessageLookupByLibrary.simpleMessage("Course textbook"),
         "courseTimes": MessageLookupByLibrary.simpleMessage("Course times"),
@@ -160,8 +168,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "importCourse": MessageLookupByLibrary.simpleMessage("Import course"),
         "importCourseWarning": MessageLookupByLibrary.simpleMessage(
             "Note! After importing, it will not really join the course selection, just to facilitate viewing the time and place"),
-        "informationSystem":
-            MessageLookupByLibrary.simpleMessage("Information System"),
+        "informationSystem": MessageLookupByLibrary.simpleMessage("Info"),
         "inputCourseName": MessageLookupByLibrary.simpleMessage(
             "Please enter the course number course name"),
         "instructor": MessageLookupByLibrary.simpleMessage("Instructor"),
@@ -181,6 +188,8 @@ class MessageLookup extends MessageLookupByLibrary {
             MessageLookupByLibrary.simpleMessage("Login course system..."),
         "loginCourseError":
             MessageLookupByLibrary.simpleMessage("Login course system error"),
+        "loginDescription": MessageLookupByLibrary.simpleMessage(
+            "A school affairs app specially designed for students of National Taiwan University of Science and Technology\nMoodle, semester schedule, score inquiry, school certificate, etc.\nAll kinds of services are available in one app"),
         "loginMoodle": MessageLookupByLibrary.simpleMessage("Login Moodle..."),
         "loginMoodleError":
             MessageLookupByLibrary.simpleMessage("Login Moodle error"),
@@ -191,6 +200,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "loginNTUST": MessageLookupByLibrary.simpleMessage("login NTUST"),
         "loginSave": MessageLookupByLibrary.simpleMessage(
             "Account password has been saved"),
+        "loginTitle": MessageLookupByLibrary.simpleMessage("NTUST TAT"),
         "logout": MessageLookupByLibrary.simpleMessage("Sign out"),
         "logoutWarning": MessageLookupByLibrary.simpleMessage(
             "Are you sure you want to log out? \nAll data will be cleared"),
