@@ -60,6 +60,26 @@ class S {
     );
   }
 
+  /// `NTUST TAT`
+  String get loginTitle {
+    return Intl.message(
+      'NTUST TAT',
+      name: 'loginTitle',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `A school affairs app specially designed for students of National Taiwan University of Science and Technology\nMoodle, semester schedule, score inquiry, school certificate, etc.\nAll kinds of services are available in one app`
+  String get loginDescription {
+    return Intl.message(
+      'A school affairs app specially designed for students of National Taiwan University of Science and Technology\nMoodle, semester schedule, score inquiry, school certificate, etc.\nAll kinds of services are available in one app',
+      name: 'loginDescription',
+      desc: '',
+      args: [],
+    );
+  }
+
   /// `Loading data error, some wrong data will be cleared`
   String get loadDataFail {
     return Intl.message(
@@ -1510,10 +1530,10 @@ class S {
     );
   }
 
-  /// `Information System`
+  /// `Info`
   String get informationSystem {
     return Intl.message(
-      'Information System',
+      'Info',
       name: 'informationSystem',
       desc: '',
       args: [],
@@ -1730,10 +1750,10 @@ class S {
     );
   }
 
-  /// `Privacy policy`
+  /// `Privacy Policy`
   String get PrivacyPolicy {
     return Intl.message(
-      'Privacy policy',
+      'Privacy Policy',
       name: 'PrivacyPolicy',
       desc: '',
       args: [],
@@ -1985,6 +2005,46 @@ class S {
     return Intl.message(
       'Please Enter WebMail Password',
       name: 'pleaseEnterWebMailPassword',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Enter text above to search for courses`
+  String get courseSearchHint {
+    return Intl.message(
+      'Enter text above to search for courses',
+      name: 'courseSearchHint',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `No relevant courses found`
+  String get courseSearchNotFound {
+    return Intl.message(
+      'No relevant courses found',
+      name: 'courseSearchNotFound',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `No Announcement`
+  String get announcementEmpty {
+    return Intl.message(
+      'No Announcement',
+      name: 'announcementEmpty',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Contribution to course total`
+  String get contributionCourse {
+    return Intl.message(
+      'Contribution to course total',
+      name: 'contributionCourse',
       desc: '',
       args: [],
     );
