@@ -38,10 +38,10 @@ class CourseTableTask extends CourseSystemTask<CourseTableJson> {
           bool add = false;
           for (int i = 0; i < 7; i++) {
             Day day = Day.values[i];
-            String? time = courseMainInfo.course.time[day];
+            String time = courseMainInfo.course.time[day] ?? "";
             courseInfo.main = courseMainInfo;
             add |=
-                courseTable.setCourseDetailByTimeString(day, time!, courseInfo);
+                courseTable.setCourseDetailByTimeString(day, time, courseInfo);
           }
           if (!add) {
             //代表課程沒有時間
