@@ -1,12 +1,13 @@
 import 'package:flutter_app/src/R.dart';
 import 'package:flutter_app/src/connector/moodle_connector.dart';
 import 'package:flutter_app/src/connector/moodle_webapi_connector.dart';
+import 'package:flutter_app/src/model/grade/tables.dart';
 import 'package:flutter_app/src/model/moodle_webapi/moodle_score.dart';
 import 'package:flutter_app/src/task/task.dart';
 
 import 'moodle_support_task.dart';
 
-class MoodleScoreTask extends MoodleSupportTask<List<MoodleScoreItem>> {
+class MoodleScoreTask extends MoodleSupportTask<TablesEntity> {
   final String courseId;
 
   MoodleScoreTask(this.courseId) : super("MoodleScoreTask", courseId) {
@@ -17,15 +18,11 @@ class MoodleScoreTask extends MoodleSupportTask<List<MoodleScoreItem>> {
   Future<TaskStatus> execute() async {
     TaskStatus status = await super.execute();
     if (status == TaskStatus.success) {
-      List<MoodleScoreItem>? value;
+      TablesEntity? value;
       super.onStart(R.current.getMoodleScore);
-      if (useMoodleWebApi) {
-        value = await MoodleWebApiConnector.getScore(findId);
-      } else {
-        value = await MoodleConnector.getScore(findId);
-      }
+      value = await MoodleWebApiConnector.getScore(findId);
       super.onEnd();
-      if (value != null && value.isNotEmpty) {
+      if (value != null) {
         result = value;
         return TaskStatus.success;
       } else {
