@@ -4,6 +4,7 @@ import 'package:flutter_app/src/connector/moodle_webapi_connector.dart';
 import 'package:flutter_app/src/model/course/course_class_json.dart';
 import 'package:flutter_app/src/model/course_table/course_table_json.dart';
 import 'package:flutter_app/src/store/model.dart';
+import 'package:flutter_app/ui/components/custom_appbar.dart';
 import 'package:flutter_app/ui/other/my_progress_dialog.dart';
 import 'package:flutter_app/ui/pages/course_data/screen/course_announcement_page.dart';
 import 'package:flutter_app/ui/pages/course_data/screen/course_directory_page.dart';
@@ -18,8 +19,8 @@ class CourseDataPage extends StatefulWidget {
   const CourseDataPage(
     this.courseInfo,
     this.index, {
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   State<StatefulWidget> createState() => _CourseDataPageState();
@@ -35,8 +36,30 @@ class _CourseDataPageState extends State<CourseDataPage>
   @override
   void initState() {
     super.initState();
-    tabPageList = TabPageList();
     testMoodleWebApi();
+
+    tabPageList = TabPageList();
+    _tabController = TabController(vsync: this, length: 3);
+    var filePage = TabPage(
+        R.current.file,
+        "img_file.svg",
+        CourseDirectoryPage(
+          widget.courseInfo,
+        ));
+    var announcementPage = TabPage(
+        R.current.announcement,
+        "img_message.svg",
+        CourseAnnouncementPage(
+          widget.courseInfo,
+        ));
+    tabPageList.add((widget.index == 0) ? filePage : announcementPage);
+    tabPageList.add((widget.index == 0) ? announcementPage : filePage);
+    tabPageList.add(TabPage(
+      R.current.score,
+      "img_education.svg",
+      CourseScorePage(widget.courseInfo),
+    ));
+    setState(() {});
   }
 
   void testMoodleWebApi() async {
@@ -49,28 +72,6 @@ class _CourseDataPageState extends State<CourseDataPage>
       MyProgressDialog.hideAllDialog();
       Model.instance.getOtherSetting().useMoodleWebApi = result;
     }
-
-    _tabController = TabController(vsync: this, length: 3);
-    var filePage = TabPage(
-        R.current.file,
-        Icons.folder,
-        CourseDirectoryPage(
-          widget.courseInfo,
-        ));
-    var announcementPage = TabPage(
-        R.current.announcement,
-        Icons.message,
-        CourseAnnouncementPage(
-          widget.courseInfo,
-        ));
-    tabPageList.add((widget.index == 0) ? filePage : announcementPage);
-    tabPageList.add((widget.index == 0) ? announcementPage : filePage);
-    tabPageList.add(TabPage(
-      R.current.score,
-      Icons.score,
-      CourseScorePage(widget.courseInfo),
-    ));
-    setState(() {});
   }
 
   @override
@@ -91,15 +92,12 @@ class _CourseDataPageState extends State<CourseDataPage>
     return DefaultTabController(
       length: tabPageList.length,
       child: Scaffold(
-        appBar: AppBar(
-          leading: BackButton(
-            onPressed: () => Get.back(),
-          ),
-          title: Text(course.name),
+        appBar: baseAppbar(
+          title: course.name,
           bottom: TabBar(
+            indicatorColor: Get.theme.indicatorColor,
             indicatorPadding: const EdgeInsets.all(0),
-            labelPadding: const EdgeInsets.all(0),
-            isScrollable: true,
+            isScrollable: false,
             controller: _tabController,
             tabs: tabPageList.getTabList(context),
             onTap: (index) {
