@@ -1,11 +1,14 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_app/src/R.dart';
 import 'package:flutter_app/src/ad/ad_manager.dart';
 import 'package:flutter_app/src/model/course_table/course_table_json.dart';
 import 'package:flutter_app/src/model/moodle_webapi/moodle_mod_forum_get_forum_discussions_paginated.dart';
 import 'package:flutter_app/src/task/moodle_webapi/moodle_course_message_task.dart';
 import 'package:flutter_app/src/task/task_flow.dart';
 import 'package:flutter_app/src/util/route_utils.dart';
-import 'package:flutter_app/ui/pages/error/error_page.dart';
+import 'package:flutter_app/ui/components/page/error_page.dart';
+import 'package:flutter_svg/flutter_svg.dart';
+import 'package:get/get.dart';
 import 'package:intl/intl.dart';
 
 class CourseAnnouncementPage extends StatefulWidget {
@@ -13,8 +16,8 @@ class CourseAnnouncementPage extends StatefulWidget {
 
   const CourseAnnouncementPage(
     this.courseInfo, {
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   State<StatefulWidget> createState() => _CourseAnnouncementPageState();
@@ -22,10 +25,6 @@ class CourseAnnouncementPage extends StatefulWidget {
 
 class _CourseAnnouncementPageState extends State<CourseAnnouncementPage>
     with AutomaticKeepAliveClientMixin {
-  @override
-  void initState() {
-    super.initState();
-  }
 
   Future<List<Discussions>?> initTask() async {
     String courseId = widget.courseInfo.main.course.id;
@@ -68,6 +67,19 @@ class _CourseAnnouncementPageState extends State<CourseAnnouncementPage>
   }
 
   Widget buildTree(List<Discussions> discussions) {
+    if(discussions.isEmpty) {
+      return Center(
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            SvgPicture.asset("assets/image/img_open_folder.svg", color: Get.iconColor, height: 72),
+            const SizedBox(height: 24),
+            Text(R.current.announcementEmpty)
+          ],
+        ),
+      );
+    }
+
     return ListView.builder(
       shrinkWrap: true,
       itemCount: discussions.length,
@@ -84,38 +96,30 @@ class _CourseAnnouncementPageState extends State<CourseAnnouncementPage>
             padding: const EdgeInsets.only(top: 10, bottom: 10),
             child: Row(
               children: [
-                const Expanded(
-                  flex: 1,
-                  child: Icon(Icons.message_outlined),
-                ),
-                Expanded(
-                  flex: 9,
-                  child: Column(
-                    children: [
-                      Row(
-                        children: [
-                          Expanded(child: Text(ap.name)),
-                          Expanded(
-                            child: Text(
-                              ap.userfullname,
-                              textAlign: TextAlign.end,
-                            ),
-                          ),
-                        ],
-                      ),
-                      Row(
-                        children: [
-                          Expanded(
-                            child: Text(
-                              formatted,
-                              textAlign: TextAlign.end,
-                            ),
-                          ),
-                        ],
-                      ),
-                    ],
+                Padding(
+                  padding: const EdgeInsets.symmetric(horizontal: 12.0),
+                  child: SvgPicture.asset(
+                    "assets/image/img_message.svg",
+                    color: Get.iconColor,
                   ),
-                )
+                ),
+                Expanded(child: Text(ap.name, style: const TextStyle(height: 1.2, overflow: TextOverflow.fade), maxLines: 2,)),
+                const SizedBox(width: 8),
+                Column(
+                  crossAxisAlignment: CrossAxisAlignment.end,
+                  children: [
+                    Text(
+                      ap.userfullname,
+                      textAlign: TextAlign.end,
+                    ),
+                    const SizedBox(height: 4),
+                    Text(
+                      formatted,
+                      textAlign: TextAlign.end,
+                    ),
+                  ],
+                ),
+                const SizedBox(width: 12),
               ],
             ),
           ),
