@@ -1,3 +1,4 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_app/src/R.dart';
 import 'package:flutter_app/src/ad/ad_manager.dart';
@@ -6,16 +7,17 @@ import 'package:flutter_app/src/model/moodle_webapi/moodle_core_course_get_conte
 import 'package:flutter_app/src/task/moodle_webapi/moodle_course_directory_task.dart';
 import 'package:flutter_app/src/task/task_flow.dart';
 import 'package:flutter_app/src/util/route_utils.dart';
+import 'package:flutter_app/ui/components/page/error_page.dart';
+import 'package:flutter_app/ui/components/tile/course_info_tile.dart';
 import 'package:flutter_app/ui/other/my_toast.dart';
-import 'package:flutter_app/ui/pages/error/error_page.dart';
 
 class CourseDirectoryPage extends StatefulWidget {
   final CourseInfoJson courseInfo;
 
   const CourseDirectoryPage(
     this.courseInfo, {
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   State<StatefulWidget> createState() => _CourseDirectoryPageState();
@@ -34,12 +36,6 @@ class _CourseDirectoryPageState extends State<CourseDirectoryPage>
     return task.result;
   }
 
-  Color getColor(int index) {
-    return (index % 2 == 1)
-        ? Theme.of(context).backgroundColor
-        : Theme.of(context).dividerColor;
-  }
-
   @override
   Widget build(BuildContext context) {
     super.build(context); //如果使用AutomaticKeepAliveClientMixin需要呼叫
@@ -56,7 +52,7 @@ class _CourseDirectoryPageState extends State<CourseDirectoryPage>
               return buildTree(snapshot.data!);
             }
           } else {
-            return const Text("");
+            return const SizedBox();
           }
         },
       ),
@@ -69,33 +65,18 @@ class _CourseDirectoryPageState extends State<CourseDirectoryPage>
       itemCount: directoryList.length,
       itemBuilder: (BuildContext context, int index) {
         var ap = directoryList[index];
-        return InkWell(
-          child: Container(
-            color: getColor(index),
-            height: 50,
-            child: Row(
-              children: [
-                const Expanded(
-                  flex: 1,
-                  child: Icon(Icons.folder_outlined),
-                ),
-                Expanded(
-                  flex: 8,
-                  child: Text(ap.name),
-                ),
-                if (ap.modules.isNotEmpty)
-                  const Icon(Icons.arrow_forward_ios_rounded, size: 16)
-              ],
-            ),
-          ),
-          onTap: () async {
-            if (ap.modules.isNotEmpty) {
+        return CourseInfoTile(
+            index: index,
+            title: ap.name,
+            img: "img_file",
+            isShowArrow: ap.modules.isNotEmpty,
+            onTap: () {
+              if (ap.modules.isEmpty) {
+                MyToast.show(R.current.nothingHere);
+                return;
+              }
               RouteUtils.toCourseInfoPage(widget.courseInfo, ap);
-            } else {
-              MyToast.show(R.current.nothingHere);
-            }
-          },
-        );
+            });
       },
     );
   }
